Feat/auth

+++ routes/api.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Classes\ApiResponseClass;
+use App\Http\Controllers\api\AuthController;
 use App\Http\Controllers\api\BeritaController;
 use App\Models\Berita;
 use Illuminate\Http\Request;
@@ -12,5 +14,13 @@ Route::get('/user', function (Request $request) {
 Route::get('/berita', [BeritaController::class, 'getAll']);
 Route::get('/berita/{id}', [BeritaController::class, 'getById']);
 Route::post('/berita', [BeritaController::class, 'store']);
-Route::put('/berita/{id}', [BeritaController::class, 'update']);
-Route::delete('/berita/{id}', [BeritaController::class, 'destroy']);
+Route::put('/berita/{id}', [BeritaController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/berita/{id}', [BeritaController::class, 'destroy'])->middleware('auth:sanctum');
+
+Route::post('/login', [AuthController::class, 'login']);
+Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
+
+//fallback route for unauthorized user
+Route::get('/unauthorize', function () {
+    return ApiResponseClass::sendError('Unauthorized', 403);
+})->name('login');
